# src/main/openapi/openapi.yaml
openapi: 3.0.0
info:
  license:
    name: MIT
  title: Petstore Open APIs
  version: 1.0.0
servers:
- url: http://petstore.swagger.io/v1
paths:
  /pets:
    get:
      operationId: listPets
      parameters:
      - description: How many items to return at one time (max 100)
        explode: true
        in: query
        name: limit
        required: false
        schema:
          format: int32
          type: integer
        style: form
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pets'
          description: A paged array of pets
          headers:
            x-next:
              description: A link to the next page of responses
              explode: false
              schema:
                type: string
              style: simple
        500:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: List all pets
      tags:
      - pets
      x-accepts: application/json
      x-tags:
      - tag: pets
    post:
      operationId: createPets
      responses:
        201:
          description: Null response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Create a pet
      tags:
      - pets
      x-accepts: application/json
      x-tags:
      - tag: pets
  /pets/{petId}:
    get:
      operationId: showPetById
      parameters:
      - description: The id of the pet to retrieve
        explode: false
        in: path
        name: petId
        required: true
        schema:
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
          description: Expected response to a valid request
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Info for a specific pet
      tags:
      - pets
      x-accepts: application/json
      x-tags:
      - tag: pets
components:
  schemas:
    Pet:
      example:
        name: name
        dateOfBirth: 2000-01-23T04:56:07.000+00:00
        id: 0
        tag: tag
      properties:
        id:
          description: Identifier of the pet
          format: int64
          type: integer
        name:
          description: the name of the pet
          type: string
        tag:
          description: Tags associated to the pets
          type: string
        dateOfBirth:
          description: date of birth of the Pet
          format: date-time
          type: string
      required:
      - id
      - name
    Pets:
      example:
        pet:
        - ""
        - ""
        - ""
        - ""
        - ""
      properties:
        pet:
          description: an array of pets
          items:
            allOf:
            - $ref: '#/components/schemas/Pet'
          maxItems: 100
          type: array
    Error:
      properties:
        code:
          description: Error Code defined for the error
          format: int32
          type: integer
        message:
          description: Message details related to the error
          type: string
      required:
      - code
      - message